Fix race condition causing threads to never start.
Suppose there are three Arachne threads A, B, and C. B is dying and C is being
born. C gets placed into slot 2, and then B dies and decrements
highestOccupiedContext. Now highestOccupiedContext points at 0, slot 1 is
UNOCCUPIED, and thread C will never run.

The solution in this commit uses two variables; a new variable to track the index of the
most recently created thread on a core and the original highestOccupiedContext.
The new variable should migrate to core, but it's possible there is a solution
using only a single variable.