Add traces around signal()
This indicates that we pay 2 cache misses and a CAS to do a signal.
The first cache miss is paid to read the wakeup flag. This wakeup flag was just
written by the target core when it woke up.
Then we pay for the CAS, which must invalidate the target core's copy.
Finally, the target core must pay another cache miss to read.

Since creation is cheaper than signal, I speculate that creation does not read
the value of the wakeup flag first. I will try to assume that the common case
is waking up a blocked thread, and first try the CAS with that value; when that
fails we can do the actual check based on the value we just read.