Remove averageSlotUtilization from core estimation
This metric was based on an instaneous estimate of threads created vs threads
completed. There are existing mechanisms in Arachne for dealing with failures
to decrease the number of cores.

Furthermore, this metric was already inaccurate because thread creations are
tracked on the creator's core rather than the target's core, and rendered
useless once we limited core estimation to shared cores.